+++ courseSystem/teachers/models.py
@@ -19,7 +19,7 @@ class Teacher(models.Model):
 class TeacherStage(models.Model):
     teacher = models.ForeignKey(to='Teacher', on_delete=models.CASCADE,verbose_name="布道师")
     stage = models.ForeignKey(to='direction.Stage', on_delete=models.CASCADE,verbose_name='阶段')
-    priority = models.IntegerField(default=0,verbose_name="权重",choices=((0.0, 0.0),(0.1, 0.1),(0.2, 0.2),(0.3, 0.3),(0.4, 0.4),(0.5, 0.5),(0.6, 0.6),(0.7, 0.7),(0.8, 0.8),(0.9, 0.9),(1.0, 1.0),))
+    priority = models.FloatField(default=0,verbose_name="权重",choices=((0.0, 0.0),(0.1, 0.1),(0.2, 0.2),(0.3, 0.3),(0.4, 0.4),(0.5, 0.5),(0.6, 0.6),(0.7, 0.7),(0.8, 0.8),(0.9, 0.9),(1.0, 1.0),))
     num = models.IntegerField(default=0,verbose_name="实施数量",editable=False)
     class Meta:
         verbose_name_plural = '布道师带课阶段表'
